requirements.txt
faiss-cpu
sentence-transformers
openai
PyPDF2
tqdm
